+++ src/modules/develop/main.js
@@ -1,7 +1,7 @@
-import yaml from 'js-yaml'
+import { safeLoad } from 'js-yaml'
 import fs from 'fs'
 
-const config = yaml.safeLoad(fs.readFileSync('config.yml')).develop
+const config = safeLoad(fs.readFileSync('config.yml')).develop
 
 const test = {
     args: '[txt...]',